Refactor selectors, do deep equal on some selector
